Fix `Cohttp_lwt_unix.Strict.get` not returning response body
* What

The `Strict.get` function has been updated to return both the HTTP
response and its body, similar to the behavior of the `Strict.call`
function.

* Why

Previously, `Strict.get` only returned the HTTP response, which was
inconsistent with the definition of a `GET` request.

* How

The internal implementation of `Strict.get` was simplified by directly
returning the result of `Strict.call`. The corresponding module
interface (`.mli`) was also updated to reflect the new return type of
`(Http.Response.t * string) Lwt.t`.